api clean : - renaming routes - game controller made - client calls corrected
